fix(service): 修复 JSON 文件生成时父目录不存在的问题
-增加了对父目录存在性的检查
- 如果父目录不存在，则不生成文件并记录日志
- 添加了单元测试来验证 JSON 文件的生成

+++ src/main/java/org/lhq/service/gen/impl/GenBookInfoJson.java
@@ -19,14 +19,19 @@ public class GenBookInfoJson implements Gen<BookInfo> {
     public void genFile(BookInfo bookInfo, File taskFile) {
         log.info("开始生成json文件");
         File parentFile = taskFile.getParentFile();
-        String path = parentFile.getPath();
-        String filePath = path + File.separator + "metadata.json";
-        try {
-            String jsonStr = Optional.ofNullable(JsonUtils.toJson(bookInfo)).orElse("");
-            Files.write(Paths.get(filePath), jsonStr.getBytes());
-            log.info("write json success {}", filePath);
-        } catch (IOException e) {
-            log.error("write json error", e);
+        if (parentFile != null && !parentFile.exists()) {
+            String path = parentFile.getPath();
+            String filePath = path + File.separator + "metadata.json";
+            try {
+                String jsonStr = Optional.ofNullable(JsonUtils.toJson(bookInfo)).orElse("");
+                Files.write(Paths.get(filePath), jsonStr.getBytes());
+                log.info("write json success {}", filePath);
+            } catch (IOException e) {
+                log.error("write json error", e);
+            }
+        }else {
+            log.info("父目录不存在");
         }
+
     }
 }
